+++ src/components/Home/MSWComponent.jsx
@@ -1,0 +1,12 @@
+'use client';
+
+import { useEffect } from 'react';
+
+export default function MSWComponent() {
+  useEffect(() => {
+    if (process.env.NEXT_PUBLIC_API_MOCKING === 'enabled') {
+      require('@/mocks/browser');
+    }
+  });
+  return null;
+}
